Capture package store from our parent Meteor process
Since we're already reaching into the Meteor ProjectContext to get the
package catalog, we can also grab the root that Meteor is using as its
package store while we're in the area.

This should fix using this package with a Meteor checkout, instead of a
released version of Meteor, as when running from a checkout, Meteor
creates a `.meteor` directory in the source tree, rather than using the
`.meteor` directory in `$HOME` or `$LOCALAPPDATA`.

This should additionally fix the integration test added in
meteor/meteor#12510.